object hangers and ERT doors update and fixes

+++ code/__HELPERS/unsorted.dm
@@ -801,6 +801,8 @@ proc/anim(turf/location as turf,target as mob|obj,a_icon,a_icon_state as text,fl
 			var/mob/mob = M
 			if(mob.client || mob.key)
 				continue
+		if(istype(M, /obj/hanger))
+			continue
 		qdel(M)
 	qdel(T)
